refactor: add readme and move files
