+++ README.md
@@ -1,1 +1,9 @@
-readme
+Precondition:
+You have to have nodejs v22
+
+For run do next steps: <br />
+
+1. npm i
+2. npm run start
+
+you will see frontend in http://localhost:8181/
