+++ src/api/personal.ts
@@ -1,5 +1,5 @@
 import { v4 as uuid } from "uuid";
-import { db } from "../db";
+import { db } from "../db/db";
 import { DbType, Events } from "../types";
 
 export const registration = async ({
